requirements.txt
pycryptodome
cffi